requirements.txt
paho-mqtt==1.6.1
pandas==2.0.0
pymongo==4.3.3
streamlit==1.17.0
scikit-learn==1.2.2
